+++ jquery.dataTables.yadcf.js
@@ -4,7 +4,7 @@
 * Yet Another DataTables Column Filter - (yadcf)
 * 
 * File:        jquery.dataTables.yadcf.js
-* Version:     0.8.9.beta.3 (grab latest stable from https://github.com/vedmack/yadcf/releases)
+* Version:     0.8.9.beta.4 (grab latest stable from https://github.com/vedmack/yadcf/releases)
 *  
 * Author:      Daniel Reznick
 * Info:        https://github.com/vedmack/yadcf
@@ -181,8 +181,10 @@
 				Required:			false
 				Type:				String
 				Default value:		undefined
-				Possible values:	chosen / select2 
+				Possible values:	chosen / select2 / custom_select
 				Description:		Turns the simple select element into Chosen / Select2 (make use of the Chosen / Select2 select jQuery plugins)
+				Note:				When using custom_select , make sure to call the initSelectPluginCustomTriggers, 
+									before calling yadcf constructor / init function
 				
 				
 * select_type_options
@@ -249,11 +251,12 @@
 									array with columns numbers 
 				Usage example:		yadcf.exResetAllFilters(oTable, [1,2]);
 				
-* setCustomSelectTriggers
- 				Description:		Allows to set custom select jquery plugin initialize and refresh method. Jquery object will be passed to the user defined function to initialize and refresh the plugin.
-				Arguments:			initializeMethod : function which can initialize the plugin	
-									refreshMethod : function which can refresh the values of plugin.
-				Usage example:		yadcf.setCustomSelectTriggers(function(obj){obj.multiselect({});}, function(obj){obj.multiselect("refresh")});
+* initSelectPluginCustomTriggers
+				Description:		Allows to set any select jquery plugin initialize and refresh functions. jQuery selector will be passed to the user defined function to initialize and refresh the plugin.
+									Great for integrating any jquey select plugin  (Selectize / MultiSelect / etc)
+				Arguments:			initFunc  : function which will initialize the plugin	
+									refreshFunc : function that will refresh the plugin.
+				Usage example:		yadcf.initSelectPluginCustomTriggers(function($filterSelector){$filterSelector.multiselect({});}, function($filterSelector){$filterSelector.multiselect("refresh")});
 									
 *
 *				
@@ -330,10 +333,9 @@ var yadcf = (function ($) {
 		exFilterColumnQueue = [],
 		yadcfDelay,
 		reA = /[^a-zA-Z]/g,
-		reN = /[^0-9]/g;
-	
-	var customSelectInitializeMethod;
-	var customSelectRefreshMethod;
+		reN = /[^0-9]/g,
+		selectElementCustomInitFunc,
+		selectElementCustomRefreshFunc;
 
 	//From ColReorder (SpryMedia Ltd (www.sprymedia.co.uk))
 	function getSettingsObjFromTable(dt) {
@@ -553,6 +555,36 @@ var yadcf = (function ($) {
 		};
 	}());
 
+	function initializeSelectPlugin(selectType, $selectObject, select_type_options) {
+		if (selectType === 'chosen') {
+			$selectObject.chosen(select_type_options);
+			$selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
+		} else if (selectType === 'select2') {
+			$selectObject.select2(select_type_options);
+			if ($selectObject.next().hasClass('select2-container')) {
+				$selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
+			}
+		} else if (selectType === 'custom_select') {
+			selectElementCustomInitFunc($selectObject);
+			$selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
+		}
+	}
+
+	function refreshSelectPlugin(selectType, $selectObject, val) {
+		if (selectType === 'chosen') {
+			$selectObject.trigger("chosen:updated");
+		} else if (selectType === 'select2') {
+			$selectObject.select2('val', val);
+		} else if (selectType === 'custom_select') {
+			selectElementCustomRefreshFunc($selectObject);
+		}
+	}
+
+	function initSelectPluginCustomTriggers(initFunc, refreshFunc) {
+		selectElementCustomInitFunc = initFunc;
+		selectElementCustomRefreshFunc = refreshFunc;
+	}
+
 	//Used by exFilterColumn for translating readable search value into proper search string for datatables filtering
 	function yadcfMatchFilterString(table_arg, column_number, selected_value, filter_match_mode, multiple) {
 		var case_insensitive = yadcf.getOptions(table_arg.selector)[column_number].case_insensitive,
@@ -639,8 +671,8 @@ var yadcf = (function ($) {
 			}
 			$("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number).val('-1').focus();
 			$("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number).removeClass("inuse");
-			
-			refreshSelectBox(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), '-1')
+
+			refreshSelectPlugin(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), '-1');
 
 		} else {
 			$("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number).addClass("inuse");
@@ -695,8 +727,8 @@ var yadcf = (function ($) {
 			$(document).data("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number + "_val", "-1");
 			oTable.fnFilter("", column_number_filter);
 			resetIApiIndex();
-			
-			refreshSelectBox(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), '-1')
+
+			refreshSelectPlugin(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), '-1');
 			return;
 		}
 
@@ -2048,9 +2080,9 @@ var yadcf = (function ($) {
 								$('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number).val(tmpStr).addClass("inuse");
 							}
 						}
-						
-						refreshSelectBox(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), tmpStr);
-						
+
+						refreshSelectPlugin(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), tmpStr);
+
 					} else if (columnObj.filter_type === "auto_complete") {
 						$(document).data("yadcf-filter-" + table_selector_jq_friendly + "-" + column_number, column_data);
 					}
@@ -2115,9 +2147,9 @@ var yadcf = (function ($) {
 							tmpStr = yadcfParseMatchFilter(tmpStr, getOptions(oTable.selector)[column_number].filter_match_mode);
 							$('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number).val(tmpStr).addClass("inuse");
 						}
-						
+
 						if (columnObj.select_type !== undefined) {
-							initializeSelectBox(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), columnObj.select_type_options);
+							initializeSelectPlugin(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), columnObj.select_type_options);
 						}
 
 					} else if (columnObj.filter_type === "multi_select" || columnObj.filter_type === 'multi_select_custom_func') {
@@ -2180,9 +2212,9 @@ var yadcf = (function ($) {
 						if (columnObj.filter_container_id !== undefined && columnObj.select_type_options.width === undefined) {
 							columnObj.select_type_options = $.extend(columnObj.select_type_options, {width: $(filter_selector_string).closest('#' + columnObj.filter_container_id).width() + "px"});
 						}
-						
+
 						if (columnObj.select_type !== undefined) {
-							initializeSelectBox(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), columnObj.select_type_options);
+							initializeSelectPlugin(columnObj.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number), columnObj.select_type_options);
 						}
 
 					} else if (columnObj.filter_type === "auto_complete") {
@@ -2631,8 +2663,8 @@ var yadcf = (function ($) {
 			} else {
 				tablesAsOne.search('').draw();
 			}
-			
-			refreshSelectBox(columnObj.select_type, $('#' + columnsObj.filter_container_id + ' select'), '-1');
+
+			refreshSelectPlugin(columnsObj.select_type, $('#' + columnsObj.filter_container_id + ' select'), '-1');
 			return;
 		}
 
@@ -2690,8 +2722,8 @@ var yadcf = (function ($) {
 			} else {
 				tablesAsOne.search('').draw();
 			}
-			
-			refreshSelectBox(columnObj.select_type, $('#' + columnsObj.filter_container_id + ' select'), '-1');
+
+			refreshSelectPlugin(columnsObj.select_type, $('#' + columnsObj.filter_container_id + ' select'), '-1');
 
 			return;
 		}
@@ -3120,9 +3152,9 @@ var yadcf = (function ($) {
 								options_tmp += "<option value=\"" + filterOptions.data[ii] + "\">" + filterOptions.data[ii] + "</option>";
 							}
 							$('#' + filterOptions.filter_container_id + ' select').empty().append(options_tmp);
-							
+
 							if (filterOptions.select_type !== undefined) {
-								initializeSelectBox(filterOptions.select_type, $('#' + filterOptions.filter_container_id + ' select'), filterOptions.select_type_options);
+								initializeSelectPlugin(filterOptions.select_type, $('#' + filterOptions.filter_container_id + ' select'), filterOptions.select_type_options);
 							}
 						});
 					}
@@ -3158,9 +3190,9 @@ var yadcf = (function ($) {
 						"onclick=\"yadcf.stopPropagation(event);yadcf.doFilterMultiTablesMultiSelect('" + tablesSelectors + "', event,'" + column_number_str + "','clear'); return false;\" class=\"yadcf-filter-reset-button\">" + filterOptions.filter_reset_button_text + "</button>");
 				}
 			}
-			
+
 			if (filterOptions.select_type !== undefined) {
-				initializeSelectBox(filterOptions.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + '-' + column_number_str), filterOptions.select_type_options);
+				initializeSelectPlugin(filterOptions.select_type, $("#yadcf-filter-" + table_selector_jq_friendly + '-' + column_number_str), filterOptions.select_type_options);
 			}
 			break;
 		default:
@@ -3279,7 +3311,7 @@ var yadcf = (function ($) {
 					tmpStr = yadcfMatchFilterString(table_arg, column_position, filter_value, optionsObj.filter_match_mode, false);
 					table_arg.fnSettings().aoPreSearchCols[column_position].sSearch = tmpStr;
 					if (optionsObj.select_type !== undefined) {
-						refreshSelectBox(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
+						refreshSelectPlugin(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
 					}
 					break;
 				case 'multi_select':
@@ -3287,7 +3319,7 @@ var yadcf = (function ($) {
 					tmpStr = yadcfMatchFilterString(table_arg, column_position, filter_value, optionsObj.filter_match_mode, true);
 					table_arg.fnSettings().aoPreSearchCols[column_position].sSearch = tmpStr;
 					if (optionsObj.select_type !== undefined) {
-						refreshSelectBox(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
+						refreshSelectPlugin(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
 					}
 					break;
 				case 'range_date':
@@ -3358,7 +3390,7 @@ var yadcf = (function ($) {
 						table_arg.fnSettings().aoPreSearchCols[column_position].sSearch = filter_value;
 					}
 					if (optionsObj.select_type !== undefined) {
-						refreshSelectBox(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), filter_value);
+						refreshSelectPlugin(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), filter_value);
 					}
 					saveStateSave(table_arg, column_number, table_selector_jq_friendly, filter_value, '');
 					break;
@@ -3520,7 +3552,7 @@ var yadcf = (function ($) {
 					$('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number).val('-1').removeClass('inuse');
 					table_arg.fnSettings().aoPreSearchCols[column_number].sSearch = '';
 					if (optionsObj.select_type !== undefined) {
-						refreshSelectBox(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
+						refreshSelectPlugin(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
 					}
 					break;
 				case 'auto_complete':
@@ -3535,7 +3567,7 @@ var yadcf = (function ($) {
 					$(document).data("#yadcf-filter-" + table_selector_jq_friendly + "-" + column_number + "_val", undefined);
 					table_arg.fnSettings().aoPreSearchCols[column_number].sSearch = '';
 					if (optionsObj.select_type !== undefined) {
-						refreshSelectBox(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
+						refreshSelectPlugin(optionsObj.select_type, $('#yadcf-filter-' + table_selector_jq_friendly + '-' + column_number), '-1');
 					}
 					break;
 				case 'range_date':
@@ -3627,37 +3659,6 @@ var yadcf = (function ($) {
 		}
 		exFilterColumn(table_arg, filtersValuesArr, true);
 	}
-	
-	
-	function initializeSelectBox(selectType, selectObject, select_type_options) {
-		if(selectType === 'chosen') {
-			selectObject.chosen(select_type_options);
-			selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
-		} else if(selectType == 'select2') {
-			selectObject.select2(select_type_options);
-			if (selectObject.next().hasClass('select2-container')) {
-				selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
-			}
-		} else if(selectType == 'custom_select') {
-			customSelectInitializeMethod(selectObject);
-			selectObject.next().attr("onclick", "yadcf.stopPropagation(event);").attr("onmousedown", "yadcf.stopPropagation(event);");
-		}
-	}
-
-	function refreshSelectBox(selectType, selectObject, val) {
-		if(selectType === 'chosen') {
-			selectObject.trigger("chosen:updated");
-		} else if(selectType == 'select2') {
-			selectObject.select2('val', val ? val : '-1');
-		} else if(selectType == 'custom_select') {
-			customSelectRefreshMethod(selectObject);
-		}
-	}
-	
-	function setCustomSelectTriggers(initializeMethod, refreshMethod) {
-		customSelectInitializeMethod = initializeMethod;
-		customSelectRefreshMethod = refreshMethod;
-	}
 
     return {
 		init : init,
@@ -3688,7 +3689,7 @@ var yadcf = (function ($) {
 		generateTableSelectorJQFriendlyNew: generateTableSelectorJQFriendlyNew,
 		exFilterExternallyTriggered: exFilterExternallyTriggered,
 		exResetFilters: exResetFilters,
-		setCustomSelectTriggers:setCustomSelectTriggers
+		initSelectPluginCustomTriggers: initSelectPluginCustomTriggers
     };
 
 }(jQuery));